// src/generators/go/type/type.ts
export type SupportedType =
  | 'string'
  | 'int'
  | 'int8'
  | 'int16'
  | 'int32'
  | 'int64'
  | 'uint'
  | 'uint8'
  | 'uint16'
  | 'uint32'
  | 'uint64'
  | 'float32'
  | 'float64'
  | 'bool'
  | 'byte'
  | 'rune';

export const supportedTypes: SupportedType[] = [
  'string',
  'int',
  'int8',
  'int16',
  'int32',
  'int64',
  'uint',
  'uint8',
  'uint16',
  'uint32',
  'uint64',
  'float32',
  'float64',
  'bool',
  'byte',
  'rune',
];